Update python benchmarks
