--- src/bench/bench_python/bench_write_h5.py
@@ -1,64 +1,0 @@
-import os
-import time
-import numpy as np
-import h5py
-import json
-
-
-shape = (100, 1024, 1024)
-bb = np.s_[:100, :1024, :1024]
-chunks = [(1, 512, 512),
-          (64, 64, 64)]
-
-# result_folder = './results_laptop'
-result_folder = './results'
-
-
-def single_write(data, chunk, compression):
-    opts = 4 if compression == 'gzip' else None
-    key = '%s_%s' % ('_'.join(str(cc) for cc in chunk),
-                     compression)
-    save_path = os.path.join(save_folder,
-                             'h5_%s.h5' % key)
-    t_write = time.time()
-    with h5py.File(save_path, 'w') as f_out:
-        ds = f_out.create_dataset('data',
-                                  shape=shape,
-                                  chunks=chunk,
-                                  compression=compression,
-                                  dtype='uint8',
-                                  compression_opts=opts)
-        ds[:] = data
-    t_write = time.time() - t_write
-    size = os.path.getsize(save_path)
-    size /= 1e6  # divide by 1e6 to get mbytes
-    return key, t_write, size
-
-
-def time_write_h5(data):
-    # compressors_h5 = [None, 'gzip', 'lzf']
-    compressors_h5 = [None,]
-    times = {}
-    for chunk in chunks:
-        for compression in compressors_h5:
-            print("Write", chunk, compression)
-            key, t_write, size = single_write(data, chunk, compression)
-            times[key] = (t_write, size)
-    with open(os.path.join(result_folder, 'reswrite_h5.json'), 'w') as f:
-        json.dump(times, f, indent=4, sort_keys=True)
-
-
-if __name__ == '__main__':
-    # path = '/home/consti/sampleA+_raw_automatically_realigned.h5'
-    path = '/home/papec/Work/playground/z5_tests/sampleA+_raw_automatically_realigned.h5'
-
-    save_folder = './tmp_files'
-    if not os.path.exists(save_folder):
-        os.mkdir(save_folder)
-    if not os.path.exists(result_folder):
-        os.mkdir(result_folder)
-
-    with h5py.File(path, 'r') as f:
-        data = f['data'][bb]
-
-    time_write_h5(data)
