--- src/bench/bench_python/clean_up.py
@@ -1,4 +1,0 @@
-from shutil import rmtree
-
-rmtree('./tmp_files')
-# rmtree('./results')
